Refactor modal components and remove unused code
- Refactored modal components and removed unused code for better organization and maintainability.
- Updated imports and component structure in CreateOrder and related files.
